refactor(saves): extract SlotsService class to saves/slots/ package
SlotsService owns the slot lifecycle: list, get-saves, set-active,
switch (with pre-flight readiness check + downloads or local cleanup),
migration between slots, deletion (local + server), and the first-sync
wizard (get_save_setup_info / confirm_slot_choice).

Constructor takes a SaveService back-reference (TYPE_CHECKING guarded)
for cross-cutting helpers (_is_save_sync_enabled, _get_server_device_id,
_get_rom_save_info, _find_save_files, _file_md5, _get_active_core,
_loop, _state, _log_debug, check_save_status_background,
get_save_status) plus a SyncEngine reference for _do_download_save
(used by _do_switch_downloads).

Public methods on SaveService (get_save_slots, get_slot_saves,
set_game_slot, switch_slot, is_save_tracking_configured,
get_save_setup_info, confirm_slot_choice, get_slot_delete_info,
delete_slot) remain as thin delegates so existing test and frontend
callers work unchanged. Private helpers (_merge_persisted_slots,
_check_slot_switch_readiness, _do_switch_downloads,
_delete_local_saves_for_switch, _migrate_slot_saves,
_validate_slot_operation, _delete_server_slot_saves) move with no
external callers and are removed from SaveService.

Sub-package layout follows the CLAUDE.md "init = re-export only"
convention: slots/__init__.py exposes SlotsService; the class lives
in slots/service.py. The _NO_MIGRATION sentinel moves with the class
and is re-imported in facade.py to preserve the confirm_slot_choice
delegate's default-argument identity.

Refs #257.
